Spice data

Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.

Usage

studio_album_tracks

Format

A tibble with 31 rows and 25 variables:

artist_name

chr Name

artist_id

chr Month price was recorded. Ranges from 1-12 for January - December

Source

https://github.com/rfordatascience/tidytuesday/tree/master/data/2021/2021-12-14